Buying army units Deserting soldier when food is not enough Refactoring Service layer

+++ src/main/java/com/beval/server/dto/response/CastleDTO.java
@@ -15,9 +15,9 @@ public class CastleDTO {
     private List<CastleBuildingDTO> buildings;
     private int coordinateX;
     private int coordinateY;
-    private int wood;
-    private int stone;
-    private int food;
+    private double wood;
+    private double stone;
+    private double food;
     private int armySize;
     private int citizens;
 }
